First implementation merging - unification of names

+++ PROI_PODATKI_Lib/Tax.cpp
@@ -1,12 +1,12 @@
 ﻿#include "Tax.h"
 
 
-float Vat::calculate_tax(float income) {//23%
+double Vat::calculate_tax(double income) {//23%
 
     return 1.23 * income;
 }
 
-float Pit::calculate_tax(float income)//Pierwszy próg podatkowy - do 85 528 zł.Do tej kwoty należny podatek PIT wynosi 17 % uszczuplony o kwotę zmniejszającą podatek;
+double Pit::calculate_tax(double income)//Pierwszy próg podatkowy - do 85 528 zł.Do tej kwoty należny podatek PIT wynosi 17 % uszczuplony o kwotę zmniejszającą podatek;
 //Drugi próg podatkowy - powyżej 85 528 zł, od nadwyżki ponad tę kwotę musimy zapłacić 32 % podatku dochodowego.
 {
     if (income >= 85528) {
@@ -15,17 +15,17 @@ float Pit::calculate_tax(float income)//Pierwszy próg podatkowy - do 85 528 zł
     return 1.32 * income;
 }
 
-float Cit::calculate_tax(float income)//19%
+double Cit::calculate_tax(double income)//19%
 {
     return 1.19 * income;
 }
 
-float Pon::calculate_tax(float income)//1zl za m^2/ grunt -> 0.61 m^2
+double Pon::calculate_tax(double income)//1zl za m^2/ grunt -> 0.61 m^2
 {
     return 4.0f;
 }
 
-float Zus::calculate_tax(float income)/*Ubezpieczenie emerytalne:
+double Zus::calculate_tax(double income)/*Ubezpieczenie emerytalne:
 Składka emerytalna wynosi 19,52% podstawy wymiaru składki. W przypadku zarobków w wysokości 10 000 zł miesięcznie:
 
 Składka emerytalna = 10 000 zł * 19,52% = 1 952 zł
