+++ PROI_PODATKI_Lib/Tax.h
@@ -1,34 +1,35 @@
 #pragma once
 class Tax
 {
-    float income;
+    double income;
 public:
-    virtual float calculate_tax(float inc) = 0;
+    virtual double calculate_tax(double inc) = 0;
 };
 
 // FIRMA-----
 class Vat : public Tax {
-
 public:
-    float calculate_tax(float inc);
+    //double operator()(double inc);
+    // double inComeWithVat =  Vat(income)
+    double calculate_tax(double inc);
 };
 
 class Cit : public Tax {
-    float calculate_tax(float inc);
+    double calculate_tax(double inc);
 };
 
 
 // Osoba FIzyczna-----
 class Pit : public Tax {
-    float calculate_tax(float inc);
+    double calculate_tax(double inc);
 };
 
 class Pon : public Tax { //Podatek Od Nieruchomosci
-    float calculate_tax(float inc);
+    double calculate_tax(double inc);
 };
 
 
 //FIrma i Osoba Fizyczna
 class Zus : public Tax {
-    float calculate_tax(float inc);
+    double calculate_tax(double inc);
 };
